draw a triangle
i feel like i understand pretty well. first off we need vertices for our shape, so we define a vbo and save that data into it. we also want to make sure gl knows how to handle our vbo, so we create a vao to store information on how to process our vbo. we finally need to tell gl how to process our vertex via shaders, so we define a simple vertex shader to define how the vertex should be handled, and a fragment shader to define how color should be set. linking them all together, we are able to render a triangle!