use an EBO and fix a few comments
the idea of an EBO is to define exactly how you want GL to render your coordinates, by telling it how to use the coordinates. The exact order can be whatever, its just the amount. You can think of specifying the indices for the vertex in the EBO as just  copying the coord directly, except you dont need to to save memory. I don't think there are any other benefits besides saving memory which would be noticeable in a larger application