understanding the relationship between VAO and shaders
so a VAO defines how it goes to the shader. it has nothing to do with the VBO coords at all, only how they should be interpreted. You can see I added a w to my original coords and changed the shader to use a vec4, and the vao attribs to match accordingly, without needing to modify the vbo at all