allow moving shape via shader
this is more efficient i believe than manipulating the vertex buffer directly, since we'd have to reupload all changes to the gpu, whereas with this we simply pass values into our shader which is already prepared and not required to reupload. idk the exact details but i do believe this is better. also NO USE GLPUSHMATRIX OR GLPOPMATRIX OR GLSCALE OR GLTRANSFORM