remove weird mouse stuff + rainbow via time. understand attribs better
this is really cool actually. so i understand the attrib thing so much better now, as well as understanding that a VBO does not NEED to be only coordinates, but rather can be anything for our vertex shader. its up to our shader to decide what parts of it become position coordinates, and what can be used in other ways.