focus VBO and EBO creation under VAO
this isn't good tbh. im making it seem like there is a dependance relationship between a VAO, VBO and EBO object, but that's not true. There's a dependance between an GL_ARRAY_BUFFER and a VAO, any GL_ARRAY_BUFFER. logically I should probably create a class of buffers, then upload them as a GLenum via the VAO for binding purposes.